renaming all dashboard files so we don't eat our history in the es6 conversions
