+++ app/assets/javascripts/school_administrator_dashboard/dashboard_components/DashboardHelpers.js
