+++ app/assets/javascripts/school_administrator_dashboard/dashboard_components/SchoolAdministratorDashboards.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Switch, Route} from 'react-router-dom';
 
-import DashboardOverview from './dashboard_overview.jsx';
-import SchoolwideAbsences from './absences_dashboard/schoolwide_absences.jsx';
-import SchoolwideTardies from './tardies_dashboard/schoolwide_tardies.jsx';
+import DashboardOverview from './DashboardOverview.jsx';
+import SchoolwideAbsences from './absences_dashboard/SchoolwideAbsences.jsx';
+import SchoolwideTardies from './tardies_dashboard/SchoolwideTardies.jsx';
 
 export default function SchoolAdministratorDashboards ( {serializedData} ) {
   const {students} = serializedData;
