+++ spec/javascripts/school_administrator_dashboard/SchoolTardiesDashboard.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import SchoolTardiesDashboard from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/tardies_dashboard/school_tardies_dashboard.jsx';
+import SchoolTardiesDashboard from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/tardies_dashboard/SchoolTardiesDashboard.jsx';
 import * as Data from './DashboardTestData.js';
 
 describe('SchoolTardiesDashboard', () => {
