+++ spec/javascripts/school_administrator_dashboard/StudentsTable.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import StudentsTable from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/students_table.jsx';
+import StudentsTable from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/StudentsTable.jsx';
 import * as Data from './DashboardTestData.js';
 
 describe('Dashboard Students Table', () => {
